requirements.txt
python-telegram-bot==22.6
fastapi==0.135.1
uvicorn[standard]==0.41.0
httpx==0.28.1
pydantic==2.12.5
PyYAML==6.0.3
cryptography==46.0.5
playwright>=1.50.0
apscheduler>=3.10.4
